requirements.txt
langchain>=0.1.0
langchain-experimental>=0.0.50
langchain-google-genai>=0.0.9
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
streamlit>=1.30.0
openpyxl>=3.1.0
numpy>=1.24.0
